// src/app/features/cross-chain-routing/services/cross-chain-routing-service/models/indexed-trade.interface.ts
import InstantTrade from '@app/features/instant-trade/models/instant-trade';
import BigNumber from 'bignumber.js';

export interface TradeAndToAmount {
  trade: InstantTrade | null;
  toAmount: BigNumber;
}

export interface IndexedTradeAndToAmount {
  providerIndex: number;
  tradeAndToAmount: TradeAndToAmount;
}