// src/app/features/cross-chain-routing/services/cross-chain-routing-service/celer/models/celer-swap-method.enum.ts
export enum CelerSwapMethod {
  SWAP_INCH = 'transferWithSwapInch',
  SWAP_INCH_NATIVE = 'transferWithSwapInchNative',

  SWAP_V2 = 'transferWithSwapV2',
  SWAP_V2_NATIVE = 'transferWithSwapV2Native',

  SWAP_V3 = 'transferWithSwapV3',
  SWAP_V3_NATIVE = 'transferWithSwapV3Native',

  SWAP_BRIDGE = 'bridgeWithSwap',
  SWAP_BRIDGE_NATIVE = 'bridgeWithSwapNative'
}